chain-code-tx/src/test/java/io/setl/bc/pychain/tx/create/RegisterAddressTest.java
/* <notice>
 
    SETL Blockchain
    Copyright (C) 2021 SETL Ltd
 
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU Affero General Public License, version 3, as
    published by the Free Software Foundation.
 
    This program is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU Affero General Public License for more details.
 
    You should have received a copy of the GNU Affero General Public License
    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
</notice> */
package io.setl.bc.pychain.tx.create;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

import io.setl.bc.pychain.state.tx.Hash;
import io.setl.bc.pychain.state.tx.RegisterAddressTx;
import io.setl.common.CommonPy.TxType;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;

public class RegisterAddressTest {

  @Test
  public void getTxType() {

    RegisterAddress thisClass = new RegisterAddress();

    assertEquals(TxType.REGISTER_ADDRESS, thisClass.getTxType());
  }


  @Test
  public void registerAddressUnsigned() throws Exception {

    int chainID = 20;
    long nonce = 123456;
    String fromPubKey = "demoPublicKey";
    String fromAddress = "demoAddress";
    String toAddress = "demoAddressToo";
    String metadata = "Meta";
    String poa = "";

    String goodHash = "91a3c5056fe61718eab80a83bb017c2dddc063a406010c5462ccae10f066af0c";

    long timestamp = 1502214664L;

    RegisterAddressTx rVal = RegisterAddress.registerAddressUnsigned(
        chainID,
        nonce,
        fromPubKey,
        fromAddress,
        "",
        metadata,
        poa);

    rVal.setTimestamp(timestamp); // Test value

    rVal.setHash(Hash.computeHash(rVal));

    assertEquals("", rVal.getHash(), goodHash);

    RegisterAddressTx txCopy = new RegisterAddressTx(rVal);
    txCopy.setHash(Hash.computeHash(txCopy));

    assertEquals("", txCopy.getHash(), goodHash);

    rVal.setHash(null);
    RegisterAddress thisClass = new RegisterAddress(rVal);

    txCopy = thisClass.create();
    txCopy.setHash(Hash.computeHash(txCopy));

    assertEquals(txCopy.getHash(), goodHash);

    txCopy = RegisterAddressTx.decodeTX(txCopy.encodeTx());
    txCopy.setHash(Hash.computeHash(txCopy));

    assertEquals("", txCopy.getHash(), goodHash);

    timestamp = 1212121212L;

    rVal = RegisterAddress.registerAddressUnsigned(
        chainID,
        nonce,
        fromPubKey,
        fromAddress,
        "",
        metadata,
        poa);

    rVal.setTimestamp(timestamp); // Test value

    rVal.setHash(Hash.computeHash(rVal));

    assertFalse(rVal.getHash().equals(goodHash));

  }


  @Before
  public void setUp() throws Exception {

  }


  @After
  public void tearDown() throws Exception {

  }


}